Refactor out getPageGroup function into its own component
